Move Slate\Standards to Slate\SBG
