+++ php-classes/Slate/SBG/PromptsRequestHandler.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Slate\Standards;
+namespace Slate\SBG;
 
 class PromptsRequestHandler extends \RecordsRequestHandler
 {
