+++ php-classes/Slate/SBG/WorksheetAssignmentsRequestHandler.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Slate\Standards;
+namespace Slate\SBG;
 
 use Slate\Term;
